+++ resources/listen.rb
@@ -1,4 +1,4 @@
-property :mode, String, equal_to: %w(http tcp)
+property :mode, String, equal_to: %w(http tcp health)
 property :bind, [String, Hash], default: '0.0.0.0:80'
 property :maxconn, Integer
 property :stats, Hash