+++ test/integration/config_1_userlist/example_spec.rb
@@ -9,22 +9,58 @@ describe directory '/etc/haproxy' do
   it { should exist }
 end
 
+cfg_content = [
+  'global',
+  '  user haproxy',
+  '  group haproxy',
+  '  log /dev/log local0',
+  '  log-tag WARDEN',
+  '  chroot /var/lib/haproxy',
+  '  daemon',
+  '  quiet',
+  '  stats socket /var/lib/haproxy/stats level admin',
+  '  maxconn 256',
+  '  pidfile /var/run/haproxy\.pid',
+  '  tune\.bufsize 262144',
+  '',
+  '',
+  'defaults',
+  '  timeout connect 5000ms',
+  '  timeout client 5000ms',
+  '  timeout server 5000ms',
+  '  log global',
+  '  mode http',
+  '  balance roundrobin',
+  '  option httplog',
+  '  option dontlognull',
+  '  option redispatch',
+  '  option tcplog',
+  '  retries 5',
+  '  stats uri /haproxy-status',
+  '',
+  '',
+  'userlist mylist',
+  '  group G1 users tiger,scott',
+  '  group G2 users xdb,scott',
+  '  user tiger password \$6\$k6y3o\.eP\$JlKBx9za9667qe4\(\.\.\.\)xHSwRv6J\.C0/D7cV91',
+  '  user scott insecure-password elgato',
+  '  user xdb insecure-password hello',
+  '',
+  '',
+  'frontend http-in',
+  '  default_backend servers',
+  '  bind \*:80',
+  '',
+  '',
+  'backend servers',
+  '  server server1 127\.0\.0\.1:8000 maxconn 32',
+]
+
 describe file('/etc/haproxy/haproxy.cfg') do
   it { should exist }
   it { should be_owned_by 'haproxy' }
   it { should be_grouped_into 'haproxy' }
-  its('content') { should match(/^  daemon$/) }
-  its('content') { should match(/^  timeout connect 5000ms$/) }
-  its('content') { should match(/^frontend http-in$/) }
-  its('content') { should match(/^  bind \*:80$/) }
-  its('content') { should match(/^backend servers$/) }
-  its('content') { should match(/^  server server1 127.0.0.1:8000 maxconn 32$/) }
-  its('content') { should match(/^userlist mylist$/) }
-  its('content') { should match(/^  group G1 users tiger,scott$/) }
-  its('content') { should match(/^  group G2 users xdb,scott$/) }
-  its('content') { should match(%r{^  user tiger password \$6\$k6y3o\.eP\$JlKBx9za9667qe4\(\.\.\.\)xHSwRv6J\.C0/D7cV91$}) }
-  its('content') { should match(/^  user scott insecure-password elgato$/) }
-  its('content') { should match(/^  user xdb insecure-password hello$/) }
+  its('content') { should match(/#{cfg_content.join('\n')}/) }
 end
 
 describe service('haproxy') do